WIP commit

+++ console/console.go
@@ -17,8 +17,8 @@ const (
 	_version     = "v0.1"
 	_logoWidth   = 57
 	_logoHeight  = 24
-	_charWidth   = int32(4)
-	_charHeight  = int32(8)
+	_charWidth   = 4
+	_charHeight  = 8
 	_cursorFlash = time.Duration(500 * time.Millisecond)
 )
